feat: update usdg price feed across instances
