+++ src/20260514_Multi_UpdateUSDGPriceFeed/AaveV3XLayer_UpdateUSDGPriceFeed_20260514.t.sol
@@ -1,0 +1,77 @@
+// SPDX-License-Identifier: MIT
+pragma solidity ^0.8.0;
+
+import {AaveV3XLayer, AaveV3XLayerAssets} from 'aave-address-book/AaveV3XLayer.sol';
+import {IAaveOracle} from 'aave-address-book/AaveV3.sol';
+
+import 'forge-std/Test.sol';
+import {ProtocolV3TestBase, ReserveConfig} from 'aave-helpers/src/ProtocolV3TestBase.sol';
+import {GovV3Helpers} from 'aave-helpers/src/GovV3Helpers.sol';
+import {IChainlinkAggregator} from 'aave-helpers/src/interfaces/IChainlinkAggregator.sol';
+import {IPriceCapAdapterStable} from '../interfaces/IPriceCapAdapterStable.sol';
+import {AaveV3XLayer_UpdateUSDGPriceFeed_20260514} from './AaveV3XLayer_UpdateUSDGPriceFeed_20260514.sol';
+
+/**
+ * @dev Test for AaveV3XLayer_UpdateUSDGPriceFeed_20260514
+ * command: FOUNDRY_PROFILE=test forge test --match-path=src/20260514_Multi_UpdateUSDGPriceFeed/AaveV3XLayer_UpdateUSDGPriceFeed_20260514.t.sol -vv
+ */
+contract AaveV3XLayer_UpdateUSDGPriceFeed_20260514_Test is ProtocolV3TestBase {
+  address internal constant USDG_PRICE_FEED = 0xe00B2732396a1f047d4A00e0165025A9cF400245;
+  address internal constant USDG_USD_CHAINLINK_FEED = 0x385C6bDDE06b0E438319bF4ddBfFe51C521ABf3D;
+
+  AaveV3XLayer_UpdateUSDGPriceFeed_20260514 internal proposal;
+
+  function setUp() public {
+    vm.createSelectFork(vm.rpcUrl('xlayer'), 63357897);
+    proposal = new AaveV3XLayer_UpdateUSDGPriceFeed_20260514();
+  }
+
+  function test_defaultProposalExecution() public {
+    defaultTest('AaveV3XLayer_UpdateUSDGPriceFeed_20260514', AaveV3XLayer.POOL, address(proposal));
+  }
+
+  function test_usdgPriceFeedMigratedToCapo() public {
+    address before = AaveV3XLayer.ORACLE.getSourceOfAsset(AaveV3XLayerAssets.USDG_UNDERLYING);
+    assertEq(before, AaveV3XLayerAssets.USDG_ORACLE, 'USDG not on the legacy feed');
+
+    GovV3Helpers.executePayload(vm, address(proposal));
+
+    assertEq(
+      AaveV3XLayer.ORACLE.getSourceOfAsset(AaveV3XLayerAssets.USDG_UNDERLYING),
+      USDG_PRICE_FEED,
+      'USDG should be repriced to the capped market feed'
+    );
+  }
+
+  function test_usdgCapoWiringAndBounds() public view {
+    IPriceCapAdapterStable capo = IPriceCapAdapterStable(USDG_PRICE_FEED);
+    assertEq(IChainlinkAggregator(USDG_PRICE_FEED).decimals(), 8, 'USDG CAPO decimals != 8');
+    assertEq(
+      capo.ASSET_TO_USD_AGGREGATOR(),
+      USDG_USD_CHAINLINK_FEED,
+      'USDG CAPO should source the USDG/USD Chainlink feed'
+    );
+    assertEq(capo.getPriceCap(), int256(1.04e8), 'USDG CAPO price cap should be 1.04');
+
+    int256 price = IChainlinkAggregator(USDG_PRICE_FEED).latestAnswer();
+    assertGt(price, int256(0.98e8), 'USDG price below expected lower bound');
+    assertLe(price, capo.getPriceCap(), 'USDG price must not exceed the cap');
+  }
+
+  function test_reservesUseNewPriceFeed() public {
+    GovV3Helpers.executePayload(vm, address(proposal));
+
+    _assertReserveUsesFeed(AaveV3XLayerAssets.USDG_UNDERLYING);
+  }
+
+  /// @dev Resolves the oracle from the pool and checks the listed reserve is priced by the new feed.
+  function _assertReserveUsesFeed(address asset) internal view {
+    assertTrue(
+      AaveV3XLayer.POOL.getReserveData(asset).aTokenAddress != address(0),
+      'asset is not a listed reserve'
+    );
+    IAaveOracle oracle = IAaveOracle(AaveV3XLayer.POOL.ADDRESSES_PROVIDER().getPriceOracle());
+    assertEq(oracle.getSourceOfAsset(asset), USDG_PRICE_FEED, 'reserve not priced by the new feed');
+    assertGt(oracle.getAssetPrice(asset), 0, 'reserve price is zero');
+  }
+}
